requirements.txt
customtkinter
pywin32; sys_platform == 'win32'
sqlite3
shutil
datetime
glob
subprocess
json
tempfile
threading
Pillow
